+++ src/routes/category-implication/[id]/+page.server.ts
@@ -1,7 +1,7 @@
 import type { EntityShort, CategoryImplicationDB } from '$lib/commons/types'
 import { batch } from '$lib/server/db.catdat'
 import { render_nested_formulas } from '$lib/server/formulas'
-import { display_implication } from '$lib/server/utils'
+import { display_implication } from '$lib/server/transforms'
 import { error } from '@sveltejs/kit'
 import sql from 'sql-template-tag'
