+++ src/routes/functor-implications/+page.server.ts
@@ -3,7 +3,7 @@ import { query } from '$lib/server/db.catdat'
 import sql from 'sql-template-tag'
 import { error } from '@sveltejs/kit'
 import type { FunctorImplicationDB } from '$lib/commons/types'
-import { display_functor_implication } from '$lib/server/utils'
+import { display_functor_implication } from '$lib/server/transforms'
 
 export const prerender = true
